+++ packages/ai-chat/src/common/parsed-chat-request.ts
@@ -20,7 +20,7 @@
 // Partially copied from https://github.com/microsoft/vscode/blob/a2cab7255c0df424027be05d58e1b7b941f4ea60/src/vs/workbench/contrib/chat/common/chatParserTypes.ts
 // Partially copied from https://github.com/microsoft/vscode/blob/a2cab7255c0df424027be05d58e1b7b941f4ea60/src/vs/editor/common/core/offsetRange.ts
 
-import { AIVariable, ResolvedAIVariable, ToolRequest, toolRequestToPromptText } from '@theia/ai-core';
+import { ResolvedAIVariable, ToolRequest, toolRequestToPromptText } from '@theia/ai-core';
 import { ChatRequest } from './chat-model';
 
 export const chatVariableLeader = '#';
@@ -41,7 +41,7 @@ export interface ParsedChatRequest {
     readonly request: ChatRequest;
     readonly parts: ParsedChatRequestPart[];
     readonly toolRequests: Map<string, ToolRequest>;
-    readonly variables: Map<string, AIVariable>;
+    readonly variables: ResolvedAIVariable[];
 }
 
 export interface ParsedChatRequestPart {